services/pre-traitement/requirements.txt
kafka-python
pandas
minio
rasterio
numpy
Pillow
py-eureka-client
